% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{builder}
\alias{builder}
\title{Builder}
\usage{
builder(d.set, normalData = 2000, percent = 80)
}
\arguments{
\item{normalData}{An integer}

\item{trainingData}{A data frame that contains attack strings, types, and
labels, followed by columns with any number of anomaly features for building
the model - must have at least 2 features to run.}
}
