% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{changesim}
\alias{changesim}
\title{Changepoint Example Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 450 rows and 20 columns.
}
\usage{
changesim
}
\description{
This dataset contains a simulated multivariate time series with two
changepoints at time point 150 and 300. The dimension of the data is T=450 and p=20.
}
\keyword{datasets}
