% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_dist.R
\name{char_dist}
\alias{char_dist}
\title{Distribution atttributes}
\usage{
char_dist(args)
}
\arguments{
\item{args}{A character vector of arguments to be sent
to determine the ICD.}
}
\value{
A matrix with 6 columns, one for each \code{dist_i}
 value of the ICD, and with the number of rows equal to the
 length of \code{args}.
}
\description{
A function that provides a wrapped around
 \code{charDist} to handle vectors of arguments supplied.
}
