% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_read_piwick.R
\name{complete_time}
\alias{complete_time}
\title{A function to retrieve all the "missing" dates and sets the number of visits to 0.}
\usage{
complete_time(aggregated_frequent)
}
\description{
A function to retrieve all the "missing" dates and sets the number of visits to 0.
}
