% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_urls.R
\name{correct_urls}
\alias{correct_urls}
\title{Extract the canonical URLs}
\usage{
correct_urls(urls)
}
\arguments{
\item{urls}{a vector of urls}
}
\value{
a vector of canonical urls
}
\description{
Extract the canonical URLs
}
