% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_anomaly_detection.R
\name{detection_anomalies_rcs}
\alias{detection_anomalies_rcs}
\title{A Function to detect the anomalies using restricted cubic square regression (from RMS)}
\usage{
detection_anomalies_rcs(sample_complete, knots = 3)
}
\arguments{
\item{details}{While you can use a different number of knots, 3 is generally a good fit.}
}
\description{
A Function to detect the anomalies using restricted cubic square regression (from RMS)
}
