% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_csv}
\alias{export_csv}
\title{export from DBs to CSV}
\usage{
export_csv(conn, table, output_path = paste0("/data/detectR/dumps/", table,
  ".csv"), chunk_length = 1e+06)
}
\arguments{
\item{conn}{a connection to the database from which the data is retrieved}

\item{table}{the table name from which the data is retrieved}

\item{output_path}{the path where the dump will be saved}
}
\description{
export from DBs to CSV
}
