% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_urls.R
\name{fix_revues_org}
\alias{fix_revues_org}
\title{Replaces revues.org by journals.opendition.org}
\usage{
fix_revues_org(urls)
}
\arguments{
\item{urls}{a vector of urls}
}
\value{
a vector of urls
}
\description{
Replaces revues.org by journals.opendition.org
}
