% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_read_piwick.R
\name{get_aggregated_piwick}
\alias{get_aggregated_piwick}
\title{Function to set up a logs_dataset object.}
\usage{
get_aggregated_piwick(aggregated_piwick_file, absolute_days = 10,
  relative_days = 0.9)
}
\description{
Contains:
\itemize{
\item The original data
\item The dataset of "frequent" visits (with number of days higher than threshold) to be used on standard modeling with rms
\item The dataset of urls with "seldom" visits (with number of days lower than threshold) to be used on Markov analysis, as the data entries are to low to allow for standard modeling.
}
}
