% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aggregated_visits.R
\name{get_aggregated_visites}
\alias{get_aggregated_visites}
\title{Function to set up a logs_dataset object.}
\usage{
get_aggregated_visites(tbl_visites, absolute_days = 10, relative_days = 0.9)
}
\arguments{
\item{tbl_visites}{a raw table with the visits per day per url}

\item{absolute_days}{the number of days a url should have at least one visit to be taken into account}

\item{relative_days}{the proportion of days a url should have at least one visit to be taken into account}
}
\description{
Contains:
\itemize{
\item The original data
\item The dataset of "frequent" visits (with number of days higher than threshold) to be used on standard modeling with rms
\item The dataset of urls with "seldom" visits (with number of days lower than threshold) to be used on Markov analysis, as the data entries are to low to allow for standard modeling.
}
}
