% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_time_series.R
\name{get_calendar_time_series}
\alias{get_calendar_time_series}
\title{A general function to get all the relevant element to decompose a calendar time serie (based on the absolute date of publication)
The main outputs are stored into a "time_serie_board" object. It includes:
\itemize{
\item The original data frame grouped by days ("all_per_day") with additional "corrected" values (for instance, substracting seasonality, trend, outliers…)
\item The seasonal week day values determined by time decomposition.
\item A cubic polynomial model for the global trend (also determined by time decomposition)
\item A set of "extreme" outliers defined by cook distance. From time to time, the website seems especially inactive (probably a maintenance issue)
}}
\usage{
get_calendar_time_series(logs_board)
}
\description{
A general function to get all the relevant element to decompose a calendar time serie (based on the absolute date of publication)
The main outputs are stored into a "time_serie_board" object. It includes:
\itemize{
\item The original data frame grouped by days ("all_per_day") with additional "corrected" values (for instance, substracting seasonality, trend, outliers…)
\item The seasonal week day values determined by time decomposition.
\item A cubic polynomial model for the global trend (also determined by time decomposition)
\item A set of "extreme" outliers defined by cook distance. From time to time, the website seems especially inactive (probably a maintenance issue)
}
}
