% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get the metadata from the solr backends from a selection of urls}
\usage{
get_metadata(urls, fields = c("id", "url", "naked_titre"),
  sleep_time = 0.05, step = 500,
  backend = "http://147.94.102.65:8983/solr/documents/select")
}
\arguments{
\item{urls}{a vector of urls}

\item{fields}{the fields to get (vector)}

\item{sleep_time}{how long (in seconds) the function should wait before two API calls. Be gentle to the backend!}

\item{step}{how many values should be sent to the API at once (an integer). Advised value is 500; choosing a higher value may lead to getting errors 400.}

\item{backend}{the solr backend

value a tibble, with as many rows as there are elements in \code{urls}.}
}
\description{
Get the metadata from the solr backends from a selection of urls
}
