% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_chunk}
\alias{get_metadata_chunk}
\title{Get the metadata from the solr backends from a selection of urls (by chunk)}
\usage{
get_metadata_chunk(urls, fields = c("id", "url", "naked_titre"), backend,
  .pb = NULL)
}
\arguments{
\item{urls}{a vector of urls}

\item{fields}{the fields to get (vector)}

\item{backend}{the solr backend}

\item{.pb}{an object to be used as a progress bar

value a tibble, with as many rows as there are elements in \code{urls}.}
}
\description{
Get the metadata from the solr backends from a selection of urls (by chunk)
}
