% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_anomaly_detection.R
\name{get_rcs_anomalies}
\alias{get_rcs_anomalies}
\title{Extract the anomalies for each group (i. e. each point beyond confidence interval)}
\usage{
get_rcs_anomalies(fit, log_value_estimated, diff_days, knots)
}
\description{
Extract the anomalies for each group (i. e. each point beyond confidence interval)
}
