% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_anomaly_detection.R
\name{get_rcs_group}
\alias{get_rcs_group}
\title{Proceed with the rcs regression for each group}
\usage{
get_rcs_group(diff_days, log_value_estimated, knots)
}
\description{
Proceed with the rcs regression for each group
}
