% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tbl_actions}
\alias{get_tbl_actions}
\title{get the table about actions}
\usage{
get_tbl_actions(conn, from)
}
\arguments{
\item{from}{the action ID to start from}
}
\description{
get the table about actions
}
