% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tbl_all_actions}
\alias{get_tbl_all_actions}
\title{collect all actions}
\usage{
get_tbl_all_actions(tbl_visites, tbl_actions, tbl_visites_actions)
}
\arguments{
\item{tbl_visites}{a table containing the visits}

\item{tbl_actions}{a table containing the actions}

\item{tbl_visites_actions}{a table linking visits and actions}
}
\description{
collect all actions
}
