% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tbl_visit_actions}
\alias{get_tbl_visit_actions}
\title{get the table linking visits and actions}
\usage{
get_tbl_visit_actions(conn, from)
}
\description{
get the table linking visits and actions
}
