% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tbl_visits}
\alias{get_tbl_visits}
\title{get the table about visits}
\usage{
get_tbl_visits(conn, from, to = Sys.Date() - 1)
}
\arguments{
\item{conn}{a connection to the relevant DB}

\item{from}{the date from which to extract data}

\item{to}{the date to which to extract data}
}
\description{
get the table about visits
}
