% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_time_series.R
\name{get_trend_model}
\alias{get_trend_model}
\title{a function to get a polynomial cubic model on the trend (determined by time decomposition)}
\usage{
get_trend_model(time_decomposition, all_per_day)
}
\description{
a function to get a polynomial cubic model on the trend (determined by time decomposition)
}
