% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_visits.R
\name{get_visits}
\alias{get_visits}
\title{Get number of visits for a given URL}
\usage{
get_visits(backend, visited_urls = NULL, from = Sys.Date(),
  to = Sys.Date())
}
\arguments{
\item{backend}{table (usually a SQL table) friom which the data is agregated}

\item{visited_urls}{a vector of urls}
}
\description{
Get number of visits for a given URL
}
