% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_time_series.R
\name{get_wday_values}
\alias{get_wday_values}
\title{a function to retrieve the effect of each day of the week (determined by time decomposition)}
\usage{
get_wday_values(time_decomposition, all_per_day, all_mean)
}
\description{
a function to retrieve the effect of each day of the week (determined by time decomposition)
}
