% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_anomaly_detection.R
\name{graph_anomalies_rcs}
\alias{graph_anomalies_rcs}
\title{Get a graph of the anomalies.}
\usage{
graph_anomalies_rcs(anomalies_dataset, article, show_sigma = FALSE)
}
\description{
Get a graph of the anomalies.
}
