% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdist.R
\name{icdist}
\alias{icdist}
\title{ICDist}
\usage{
icdist(argument)
}
\arguments{
\item{argument}{The argument/query supplied to the field.}
}
\value{
Returns a binary flags for the anomaly check. "1" for anomalous,
 "0" for normal.
}
\description{
A function that compares a requests character distribution with
 normal data.
}
\details{
A character distribution is created and compared against an Idealised
 Character Distribution (ICD) using a Kolmogorov–Smirnov test.
}
\keyword{anomaly}
\keyword{argument}
\keyword{query}
\keyword{request}
