% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{LRV.bartlett}
\alias{LRV.bartlett}
\title{(multivariate) Bartlett long-run variance calculation}
\usage{
LRV.bartlett(cL, q)
}
\description{
(multivariate) Bartlett long-run variance calculation
}
\keyword{internal}
