% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_token.R
\name{max_token}
\alias{max_token}
\title{Max token Score}
\usage{
max_token(token, maliciousTokens = c("x", "y", "z"))
}
\arguments{
\item{token}{A character string representing the tokenization
of a string.}
}
\description{
The following patterns map to the tokens
}
