% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkpca.R
\name{networkpca}
\alias{networkpca}
\title{If model is indicated, reduce data to components.}
\usage{
networkpca(model, Y)
}
\description{
If model is indicated, reduce data to components.
}
\keyword{internal}
