% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_creator.R
\name{nn_creator}
\alias{nn_creator}
\title{Neural Network Creator}
\usage{
nn_creator(d.features, posClass = "N", normalData = 2000, percent = 80,
  logs)
}
\arguments{
\item{normalData}{An integer}

\item{trainingData}{A data frame that contains attack strings, types, and
labels, followed by columns with any number of anomaly features for building
the model - must have at least 2 features to run.}
}
