% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readership_model_read_piwick.R
\name{read_aggregated_piwick_file}
\alias{read_aggregated_piwick_file}
\title{Functions to read and clean the aggregated piwick files.}
\usage{
read_aggregated_piwick_file(aggregated_piwick_file)
}
\description{
Functions to read and clean the aggregated piwick files.
}
