% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_args.R
\name{split_args}
\alias{split_args}
\title{Split Args}
\usage{
split_args(urls)
}
\arguments{
\item{urls}{An array of URL strings. These will be URL-encoded but httr will
decode these in the lappy function}
}
\value{
A data frame containing all of the argument names, values, and character
 lengths.
}
\description{
A function that splits URL parameters into parameter value pairs,
 and the length of the query strings.
}
