% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPCA.R
\name{testPCA}
\alias{testPCA}
\title{PCA-based test for the equality of connectivity}
\usage{
testPCA(subY1, subY2, L = 2, nlag, diagTF = TRUE)
}
\arguments{
\item{subY1}{a sample of size length*dim}

\item{subY2}{a sample of size length*dim}

\item{L}{the number of factors}

\item{nlag}{is the number of ACF lag to be used in the test, default is 2, Default is nlag = floor(N^(1/3))}

\item{diagTF}{include diagonal term of covariance matrix or not}
}
\value{
\strong{tstat} Test statistic

\strong{pval} Returns the p-value

\strong{df} The degree of freedom in PCA-best test

\strong{L} The number of factors used in the test

\strong{diagTF} If true, the diagonal entry of covarianc matrix is used in testing
}
\description{
This function performs PCA-test for testing the equality of connectivity in two fMRI signals
}
\examples{
test3 = testPCA(testsim$X, testsim$Y, L=2)
}
