% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{data}
\name{testsim}
\alias{testsim}
\title{Test Example Data}
\format{
An object of class \code{list} of length 2.
}
\usage{
testsim
}
\description{
This dataset contains a simulated multivariate time series with two different autocovariances.
It is a list data with two variables X and Y. Each multivariate time series had dimension of
T=150 and p=20
}
\keyword{datasets}
