% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_hash.R
\name{token_hash}
\alias{token_hash}
\title{Token hash}
\usage{
token_hash(elevate)
}
\arguments{
\item{elevate}{A character string which is one of;
\itemize{
  \item{\code{N} : To indicate a normal classified string.}
  \item{\code{X} : To indicate a SQLi classified string.}
  \item{\code{Y} : To indicate a XSS classified string.}
  \item{\code{Z} : To indicate a Bash classified string.}
}}
}
