% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_urls.R
\name{unfix_revues_org}
\alias{unfix_revues_org}
\title{Replaces journals.opendition.org by revues.org}
\usage{
unfix_revues_org(urls)
}
\arguments{
\item{urls}{a vector of urls}
}
\value{
a vector of urls
}
\description{
Replaces journals.opendition.org by revues.org
}
