% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_urls.R
\name{validate_url}
\alias{validate_url}
\title{Validate that the url is a valid url}
\usage{
validate_url(urls, logical = FALSE)
}
\arguments{
\item{urls}{a vector of urls}

\item{logical}{should the function return a vector of logical values?}
}
\value{
a vector containing only the valid urls, or a vector of logical values
}
\description{
Validate that the url is a valid url
}
