% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{extract_and_load}
\alias{extract_and_load}
\title{Extract from a given DB and load into another DB}
\usage{
extract_and_load(conn1, conn2, from, progress = TRUE)
}
\arguments{
\item{conn1}{a connection to the first database.}

\item{conn2}{a connection to the second database.}

\item{progress}{whether a progress bar should be shown.}
}
\description{
Extract from a given DB and load into another DB
}
