% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tbl_visits}
\alias{get_tbl_visits}
\title{get the table about visits}
\usage{
get_tbl_visits(conn, from)
}
\arguments{
\item{conn}{a connection to the relevant DB}

\item{from}{the visit ID to start from}
}
\description{
get the table about visits
}
